separated PlayerList from GameRoom
