+++ zhuo_hei_cha_frontend/.vscode/launch.json
@@ -1,0 +1,15 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+      {
+        "type": "chrome",
+        "request": "launch",
+        "name": "Debug React Run",
+        "url": "http://localhost:3000",
+        "webRoot": "${workspaceFolder}/src"
+    }
+    ]
+}
